+++ release-notes.md
@@ -1,2 +1,29 @@
-No changes
+### New Endpoints: None
+-----------------------
+
+### Deleted Endpoints: None
+---------------------------
+
+### Modified Endpoints: 4
+-------------------------
+GET /api/v1/cloudcredentials/flavors/{cloudId}
+- New query param: GpuCount
+- Deleted query param: EndGpuCount
+- Deleted query param: StartGpuCount
+
+GET /api/v1/flavors/aws/{cloudId}
+- New query param: GpuCount
+- Deleted query param: EndGpuCount
+- Deleted query param: StartGpuCount
+
+GET /api/v1/flavors/azure/{cloudId}
+- New query param: GpuCount
+- Deleted query param: EndGpuCount
+- Deleted query param: StartGpuCount
+
+GET /api/v1/flavors/google/{cloudId}
+- New query param: GpuCount
+- Deleted query param: EndGpuCount
+- Deleted query param: StartGpuCount
+